@@ -6,7 +6,7 @@
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
-    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "lint": "eslint --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview"
   },
   "dependencies": {
@@ -28,6 +28,7 @@
     "@tanstack/react-query": "^5.37.1",
     "@tanstack/react-router": "^1.33.4",
     "@tanstack/react-table": "^8.17.3",
+    "@types/eslint__js": "^8.42.3",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "date-fns": "^3.6.0",
@@ -53,13 +54,12 @@
     "@types/node": "^20.12.12",
     "@types/react": "^18.2.66",
     "@types/react-dom": "^18.2.22",
-    "@typescript-eslint/eslint-plugin": "^7.2.0",
     "@typescript-eslint/parser": "^7.2.0",
     "@vitejs/plugin-react-swc": "^3.5.0",
     "autoprefixer": "^10.4.19",
-    "eslint": "9.x",
+    "eslint": "^9.8.0",
     "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-react": "^7.34.1",
+    "eslint-plugin-react": "^7.35.0",
     "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-react-refresh": "^0.4.6",
     "globals": "^15.3.0",
@@ -68,6 +68,7 @@
     "tailwindcss": "^3.4.3",
     "typescript": "^5.2.2",
     "typescript-eslint": "^7.9.0",
-    "vite": "^5.2.0"
+    "vite": "^5.2.0",
+    "vite-plugin-eslint": "^1.8.1"
   }
 }
