@@ -1,44 +1,57 @@
-import { fixupConfigRules } from '@eslint/compat';
-import pluginJs from '@eslint/js';
-import * as reactQuery from '@tanstack/eslint-plugin-query';
-import pluginImport from 'eslint-plugin-import';
-import pluginReactConfig from 'eslint-plugin-react/configs/recommended.js';
-import * as reactHooks from 'eslint-plugin-react-hooks';
 import globals from 'globals';
-import tseslint from 'typescript-eslint';
+import eslint from '@eslint/js';
+import ts from 'typescript-eslint';
+import reactRefresh from 'eslint-plugin-react-refresh';
+import reactHooks from 'eslint-plugin-react-hooks';
+import importPlugin from 'eslint-plugin-import';
+import reactPlugin from 'eslint-plugin-react';
+
+import * as tanstackQuery from '@tanstack/eslint-plugin-query';
 
 export default [
-  { languageOptions: { globals: globals.browser } },
-  pluginJs.configs.recommended,
-  ...tseslint.configs.recommended,
-  ...fixupConfigRules(pluginReactConfig),
+  eslint.configs.recommended,
+  ...ts.configs.recommended,
   {
+    ignores: ['dist/**/*', 'node_modules/**/*', 'eslint.config.js'],
+  },
+  {
+    files: ['*.ts', '*.tsx'],
+    languageOptions: {
+      ecmaVersion: 2023,
+      sourceType: 'module',
+      globals: {
+        ...globals.browser,
+      },
+      parserOptions: {
+        ecmaFeatures: {
+          jsx: true,
+        },
+      },
+    },
     plugins: {
-      '@tanstack/query': reactQuery,
-      import: { rules: pluginImport.rules },
-      'react-hooks/recommended': reactHooks,
+      react: reactPlugin,
+      'react-hooks': reactHooks,
+      '@tanstack/query': {
+        rules: tanstackQuery.rules,
+      },
+      import: { rules: importPlugin.rules },
     },
     rules: {
-      '@tanstack/query/exhaustive-deps': 'error',
-      '@tanstack/query/no-rest-destructuring': 'warn',
-      '@tanstack/query/stable-query-client': 'error',
-      'import/exports-last': 'off',
-      'import/no-useless-path-segments': [
-        'error',
-        {
-          noUselessIndex: true,
-        },
-      ],
-      'import/order': [
-        'error',
-        {
-          alphabetize: { caseInsensitive: true, order: 'asc' },
-          groups: ['index', 'sibling', 'parent', 'internal', 'external', 'builtin', 'object', 'type'],
-          'newlines-between': 'always-and-inside-groups',
-        },
-      ],
-      '@typescript-eslint/no-explicit-any': 'off',
-      'react/react-in-jsx-scope': 'off',
+      ...reactPlugin.configs.flat.recommended.rules,
+      ...tanstackQuery.configs.recommended.rules,
+    },
+    settings: {
+      react: {
+        version: 'detect',
+      },
+    },
+  },
+  {
+    plugins: {
+      'react-refresh': reactRefresh,
+    },
+    rules: {
+      'react-refresh/only-export-components': ['warn', { allowConstantExport: true }],
     },
   },
 ];
